Avoid use of ENV variables to select Gtk version in test tasks
Using an ENV variable had several problems:
- A separate task had to be defined that just set an ENV var. This task
  was useless on its own.
- The gtk3 task had to be run first. This could have been solved by
  adding another task to set GTK_VERSION to 3

With this change, the problem of specifying the Gtk version is solved
differently and more directly: Requiring the correct loader up front
makes sure the right version is loaded. In the test helper, the ENV var
mechanism is only triggered if no version has been loaded yet. To
ensure the developer that the right version has been loaded, the Gtk
version used is printed.